extract construct-measure-fragment function
